+++ frontend/components/base/base-render-markdown/renderMarkdown.base.component.spec.js
@@ -60,4 +60,32 @@ describe("RenderMarkdownBaseComponent", () => {
 </div>`
     );
   });
+
+  it("open in other window if the node is a link", () => {
+    const wrapper = shallowMount(RenderMarkdownBaseComponent, {
+      ...options,
+      propsData: {
+        markdown: `[example](https://example.com)`,
+      },
+    });
+    expect(wrapper.html()).toBe(
+      `<div class="markdown-render">
+  <p><a target="_blank" href="https://example.com">example</a></p>
+</div>`
+    );
+  });
+
+  it("open in other window if the node already hace target blank", () => {
+    const wrapper = shallowMount(RenderMarkdownBaseComponent, {
+      ...options,
+      propsData: {
+        markdown: `<a href="https://example.com" target="_blank">example</a>`,
+      },
+    });
+    expect(wrapper.html()).toBe(
+      `<div class="markdown-render">
+  <p><a target="_blank" href="https://example.com">example</a></p>
+</div>`
+    );
+  });
 });