+++ frontend/components/feedback-task/container/questions/form/ranking/drag-and-drop-selection/DndSelection.component.vue
@@ -23,15 +23,16 @@
           v-text="item.text"
           :id="`${item.value}-span`"
         />
+
         <BaseTooltip
-          v-if="findRankSuggestion(item.value)"
+          v-if="findRankSuggestion(item)"
           class="draggable__suggestion"
-          :text="`<img src=${suggestionIcon} /> ${$t(
+          :text="`<img src='/icons/suggestion.svg' /> ${$t(
             'suggestion.suggested-rank'
           )}`"
           minimalist
         >
-          <span v-text="findRankSuggestion(item.value).rank" />
+          <span v-text="findRankSuggestion(item).rank" />
         </BaseTooltip>
       </div>
     </draggable>
@@ -69,12 +70,17 @@
               v-text="item.text"
               :id="`${item.value}-span`"
             />
-            <span
-              :title="$t('suggestion.suggested-rank')"
+
+            <BaseTooltip
+              v-if="findRankSuggestion(item)"
               class="draggable__suggestion"
-              v-if="findRankSuggestion(item.value)"
-              >{{ findRankSuggestion(item.value).rank }}</span
+              :text="`<img src='/icons/suggestion.svg' /> ${$t(
+                'suggestion.suggested-rank'
+              )}`"
+              minimalist
             >
+              <span v-text="findRankSuggestion(item).rank" />
+            </BaseTooltip>
           </div>
         </draggable>
       </div>
@@ -84,7 +90,6 @@
 
 <script>
 import "assets/icons/draggable";
-import suggestionIcon from "@/static/icons/suggestion.svg";
 
 export default {
   name: "DndSelectionComponent",
@@ -93,8 +98,8 @@ export default {
       type: Object,
       required: true,
     },
-    suggestions: {
-      type: Array,
+    suggestion: {
+      type: Object,
     },
     isFocused: {
       type: Boolean,
@@ -105,7 +110,6 @@ export default {
     return {
       timer: null,
       keyCode: "",
-      suggestionIcon,
     };
   },
   watch: {
@@ -204,8 +208,8 @@ export default {
     onFocus() {
       this.$emit("on-focus");
     },
-    findRankSuggestion(value) {
-      return this.suggestions?.find((suggestion) => suggestion.value == value);
+    findRankSuggestion(item) {
+      return this.suggestion?.getSuggestion(item);
     },
   },
 };
@@ -285,6 +289,11 @@ $max-visible-card-items: 12;
       box-shadow: none;
       &:hover {
         box-shadow: none;
+      }
+    }
+    &.sortable-chosen {
+      .tooltip {
+        display: none;
       }
     }
     &--unranked {