+++ frontend/components/feedback-task/container/questions/form/shared-components/label-selection/labelSelection.component.spec.js
@@ -7,6 +7,7 @@ const options = {
   propsData: {
     componentId: `componentId`,
     options: [],
+    maxOptionsToShowBeforeCollapse: 0,
   },
 };
 beforeEach(() => {
@@ -15,6 +16,7 @@ beforeEach(() => {
 
 afterEach(() => {
   wrapper.destroy();
+  window.questionSettings["componentId"].isExpandedLabelQuestions = false;
 });
 
 describe("LabelSelectionComponent in Single Selection mode", () => {
@@ -26,7 +28,6 @@ describe("LabelSelectionComponent in Single Selection mode", () => {
     expect(wrapper.vm.searchInput).toBe(``);
     expect(wrapper.vm.showSearch).toBe(false);
     expect(wrapper.vm.showCollapseButton).toBe(false);
-    expect(wrapper.vm.maxOptionsToShowBeforeCollapse).toBe(-1);
     expect(wrapper.vm.searchRef).toBe(`componentIdSearchFilterRef`);
 
     const searchWrapper = wrapper.findComponent({
@@ -53,6 +54,7 @@ describe("LabelSelectionComponent in Single Selection mode", () => {
 
     await wrapper.setProps({
       options,
+      maxOptionsToShowBeforeCollapse: 1,
     });
 
     // by default it's a single selection
@@ -97,6 +99,7 @@ describe("LabelSelectionComponent in Single Selection mode", () => {
 
     await wrapper.setProps({
       options,
+      maxOptionsToShowBeforeCollapse: 2,
     });
 
     // by default it's a single selection
@@ -150,6 +153,7 @@ describe("LabelSelectionComponent in Single Selection mode", () => {
 
     await wrapper.setProps({
       options,
+      maxOptionsToShowBeforeCollapse: 3,
     });
 
     // by default it's a single selection
@@ -159,7 +163,7 @@ describe("LabelSelectionComponent in Single Selection mode", () => {
     const searchWrapper = wrapper.findComponent({
       name: "SearchLabelComponent",
     });
-    expect(searchWrapper.exists()).toBe(true);
+    expect(searchWrapper.exists()).toBe(false);
 
     // by default there is no collapse button
     const showLessButtonWrapper = wrapper.findComponent({
@@ -206,6 +210,7 @@ describe("LabelSelectionComponent in Single Selection mode", () => {
 
     await wrapper.setProps({
       options,
+      maxOptionsToShowBeforeCollapse: 3,
     });
 
     // by default it's a single selection
@@ -279,6 +284,7 @@ describe("LabelSelectionComponent in Single Selection mode", () => {
 
     await wrapper.setProps({
       options,
+      maxOptionsToShowBeforeCollapse: 3,
     });
 
     // by default it's a single selection
@@ -352,6 +358,7 @@ describe("LabelSelectionComponent in Single Selection mode", () => {
 
     await wrapper.setProps({
       options,
+      maxOptionsToShowBeforeCollapse: 3,
       showSearch: true,
     });
 
@@ -359,7 +366,7 @@ describe("LabelSelectionComponent in Single Selection mode", () => {
     expect(wrapper.vm.multiple).toBe(false);
 
     const searchWrapper = wrapper.findComponent({ ref: "searchComponentRef" });
-    expect(searchWrapper.exists()).toBe(true);
+    expect(searchWrapper.exists()).toBe(false);
 
     await wrapper.setData({ searchInput: "" });
     expect(wrapper.vm.filteredOptions).toStrictEqual([
@@ -542,6 +549,7 @@ describe("LabelSelectionComponent in Multi Selection mode", () => {
 
     await wrapper.setProps({
       options,
+      maxOptionsToShowBeforeCollapse: 3,
       multiple: true,
     });
 
@@ -552,7 +560,7 @@ describe("LabelSelectionComponent in Multi Selection mode", () => {
     const searchWrapper = wrapper.findComponent({
       name: "SearchLabelComponent",
     });
-    expect(searchWrapper.exists()).toBe(true);
+    expect(searchWrapper.exists()).toBe(false);
 
     // by default there is no collapse button
     const showLessButtonWrapper = wrapper.findComponent({
@@ -599,6 +607,7 @@ describe("LabelSelectionComponent in Multi Selection mode", () => {
 
     await wrapper.setProps({
       options,
+      maxOptionsToShowBeforeCollapse: 3,
       multiple: true,
     });
 
@@ -673,6 +682,7 @@ describe("LabelSelectionComponent in Multi Selection mode", () => {
 
     await wrapper.setProps({
       options,
+      maxOptionsToShowBeforeCollapse: 3,
       multiple: true,
     });
 
@@ -747,6 +757,7 @@ describe("LabelSelectionComponent in Multi Selection mode", () => {
 
     await wrapper.setProps({
       options,
+      maxOptionsToShowBeforeCollapse: 3,
       multiple: true,
       showSearch: true,
     });
@@ -755,7 +766,7 @@ describe("LabelSelectionComponent in Multi Selection mode", () => {
     expect(wrapper.vm.multiple).toBe(true);
 
     const searchWrapper = wrapper.findComponent({ ref: "searchComponentRef" });
-    expect(searchWrapper.exists()).toBe(true);
+    expect(searchWrapper.exists()).toBe(false);
 
     await wrapper.setData({ searchInput: "" });
     expect(wrapper.vm.filteredOptions).toStrictEqual([