+++ frontend/v1/domain/entities/question/Question.ts
@@ -1,13 +1,16 @@
 import { Answer } from "../IAnswer";
+import { Guard } from "../error";
+import { Color } from "./Color";
 import {
   QuestionAnswer,
-  QuestionType,
   TextQuestionAnswer,
   SingleLabelQuestionAnswer,
   RatingLabelQuestionAnswer,
   MultiLabelQuestionAnswer,
   RankingQuestionAnswer,
+  SpanQuestionAnswer,
 } from "./QuestionAnswer";
+import { QuestionType } from "./QuestionType";
 import { Suggestion } from "./Suggestion";
 
 interface OriginalQuestion {
@@ -18,7 +21,7 @@ interface OriginalQuestion {
 
 export class Question {
   public answer: QuestionAnswer;
-  private suggestion: Suggestion;
+  public suggestion: Suggestion;
   private original: OriginalQuestion;
 
   constructor(
@@ -33,6 +36,7 @@ export class Question {
     this.description = description;
     this.title = title;
 
+    this.initialize();
     this.initializeAnswers();
     this.initializeOriginal();
   }
@@ -64,31 +68,31 @@ export class Question {
   }
 
   public get type(): QuestionType {
-    return this.settings.type.toLowerCase();
+    return QuestionType.from(this.settings.type);
   }
 
   public get isRankingType(): boolean {
-    return this.type === "ranking";
+    return this.type.isRankingType;
   }
 
   public get isMultiLabelType(): boolean {
-    return this.type === "multi_label_selection";
+    return this.type.isMultiLabelType;
   }
 
   public get isSingleLabelType(): boolean {
-    return this.type === "label_selection";
+    return this.type.isSingleLabelType;
   }
 
   public get isTextType(): boolean {
-    return this.type === "text";
+    return this.type.isTextType;
+  }
+
+  public get isSpanType(): boolean {
+    return this.type.isSpanType;
   }
 
   public get isRatingType(): boolean {
-    return this.type === "rating";
-  }
-
-  public get matchSuggestion(): boolean {
-    return !!this.suggestion && this.answer.matchSuggestion(this.suggestion);
+    return this.type.isRatingType;
   }
 
   public get isModified(): boolean {
@@ -154,14 +158,6 @@ export class Question {
     this.answer = questionReference.answer;
   }
 
-  responseIfUnanswered(answer: Answer) {
-    if (this.suggestion) {
-      this.answer.responseIfUnanswered(this.suggestion);
-    } else if (answer) {
-      this.answer.responseIfUnanswered(answer);
-    }
-  }
-
   response(answer: Answer) {
     if (!answer) return;
 
@@ -174,13 +170,21 @@ export class Question {
     this.suggestion = suggestion;
   }
 
-  initializeAnswers() {
-    this.answer = this.createEmptyAnswers();
+  reloadAnswerFromOptions() {
+    this.initializeAnswers();
   }
 
   private createEmptyAnswers(): QuestionAnswer {
     if (this.isTextType) {
       return new TextQuestionAnswer(this.type, "");
+    }
+
+    if (this.isSpanType) {
+      return new SpanQuestionAnswer(
+        this.type,
+        this.name,
+        this.settings.options
+      );
     }
 
     if (this.isRatingType) {
@@ -214,6 +218,31 @@ export class Question {
         this.settings.options
       );
     }
+
+    Guard.throw(
+      `Question answer for type ${this.type} is not implemented yet.`
+    );
+  }
+
+  private initialize() {
+    if (this.settings.options && !this.settings.visible_options) {
+      this.settings.visible_options = this.settings.options.length;
+    }
+
+    if (this.isSpanType) {
+      this.settings.options = this.settings.options.map((option) => {
+        return {
+          ...option,
+          color: option.color
+            ? Color.from(option.color)
+            : Color.generate(option.value),
+        };
+      });
+    }
+  }
+
+  private initializeAnswers() {
+    this.answer = this.createEmptyAnswers();
   }
 
   private initializeOriginal() {